Seperated sending/receiving tasks and functions into its own class and file; Minor cleanup of main display code
