+++ MainDisplay.cs
@@ -9,6 +9,8 @@ namespace AT_SCC
     {                                        
         
         private readonly DisplayHelp _displayHelp;
+
+        private readonly SerialHelp _serialHelp;
         
            // declaring all variables associated with textboxes, menu/tool strip, strings, or int
         private readonly TextBox? textBoxCOM = new(), textBoxBAUD = new(), textBoxPARITY = new(), textBoxDATABITS = new(), textBoxSTOPBITS = new(), textBoxRTIMEOUT = new(), textBoxWTIMEOUT = new(), textBoxHANDSHAKE = new(), textBoxDELAY = new(), textBoxBTT = new(), textBoxMODEDISP = new(), textBoxSENDTYPE = new(), textBoxreceiveType = new();
@@ -73,32 +75,20 @@ namespace AT_SCC
 
         public readonly Dictionary<string, Parity> ParityOptions = new()
         {
-            ["None"] = Parity.None,
-            ["Even"] = Parity.Even,
-            ["Odd"] = Parity.Odd,
-            ["Mark"] = Parity.Mark,
-            ["Space"] = Parity.Space
+            ["None"] = Parity.None,["Even"] = Parity.Even, ["Odd"] = Parity.Odd,["Mark"] = Parity.Mark,["Space"] = Parity.Space
         };
 
         public readonly Dictionary<string, StopBits> StopBitOptions = new()
         {
-            ["1"] = StopBits.One,
-            ["1.5"] = StopBits.OnePointFive,
-            ["2"] = StopBits.Two
+            ["1"] = StopBits.One,["1.5"] = StopBits.OnePointFive,["2"] = StopBits.Two
         };
 
         public readonly Dictionary<string, Handshake> HandShakeOptions = new()
         {
-            ["None"] = Handshake.None,
-            ["XOnXOff"] = Handshake.XOnXOff,
-            ["Send (Rq)"] = Handshake.RequestToSend,
-            ["XOnXOff (Rq)"] = Handshake.RequestToSendXOnXOff
+            ["None"] = Handshake.None,["XOnXOff"] = Handshake.XOnXOff,["Send (Rq)"] = Handshake.RequestToSend,["XOnXOff (Rq)"] = Handshake.RequestToSendXOnXOff
         };
 
-        public readonly string[] sendDelayOptions =
-        [
-            "100","500","1000","2000","3000","4000","5000"
-        ];
+        public readonly string[] sendDelayOptions =["100","500","1000","2000","3000","4000","5000"];
 
         // DECLARING EVENTS
         private void TextBoxBTT_TextChanged(object sender, EventArgs e)     // event to adjust the TX BUFFER SIZE
@@ -132,7 +122,7 @@ namespace AT_SCC
             });
         }
 
-        private void Transmission_Click(object? sender, System.EventArgs e) // event to handle clicking the start transmission button
+        private void Transmission_Click(object? sender, EventArgs e) // event to handle clicking the start transmission button
         {
             if (transmitactive == 1)
             {
@@ -240,7 +230,7 @@ namespace AT_SCC
                         {
                             // Send the textbox contents as a string
                             var textToSend = textBox.Text;
-                            await SendStringAsync(mySerialPort, textToSend, logging_check.Checked, LogFilePath);
+                            await _serialHelp.SendStringAsync(mySerialPort, textToSend, logging_check.Checked, LogFilePath, textBoxDELAY!);
                         }
                     }
                 }
@@ -248,7 +238,7 @@ namespace AT_SCC
                 if (textBoxreceiveType?.Text == "String" && (textBoxMODEDISP?.Text == "Receive Mode" || textBoxMODEDISP?.Text == "Send and Receive"))
                 {
                     var receivedTextBox = textBoxArray[i];
-                    ReceiveStringAsync(mySerialPort, receivedTextBox, logging_check.Checked, LogFilePath); // receives the data and sets to output panel
+                    _serialHelp.ReceiveStringAsync(mySerialPort, receivedTextBox, logging_check.Checked, LogFilePath); // receives the data and sets to output panel
                     i++;
 
                 }
@@ -301,7 +291,7 @@ namespace AT_SCC
                         {
                             // Send the textbox contents as a string
                             var textToSend = textBox.Text;
-                            await SendBytesAsync(mySerialPort, textBoxSENDTYPE.Text, textBoxesPanel.Controls.OfType<TextBox>());
+                            await _serialHelp.SendBytesAsync(mySerialPort, textBoxesPanel.Controls.OfType<TextBox>(), logging_check, LogFilePath, textBoxDELAY!);
                         }
                     }
                 }
@@ -309,7 +299,7 @@ namespace AT_SCC
                 if (textBoxreceiveType?.Text == "Byte/Byte Collection" && (textBoxMODEDISP?.Text == "Receive Mode"  || textBoxMODEDISP?.Text == "Send and Receive"))
                 {
 
-                    ReceiveBytesAsync(mySerialPort, textBoxArray, i); // receives the data and sets to output panel
+                    _serialHelp.ReceiveBytesAsync(mySerialPort, textBoxArray, i, logging_check, LogFilePath); // receives the data and sets to output panel
                     i++;
 
                 }
@@ -351,7 +341,7 @@ namespace AT_SCC
                 var hexBytesList = new List<byte>();
                 foreach (var value in inputValues)
                 {
-                    if (byte.TryParse(value, System.Globalization.NumberStyles.HexNumber, null, out var hexByte))
+                    if (byte.TryParse(value, NumberStyles.HexNumber, null, out var hexByte))
                     {
                         hexBytesList.Add(hexByte);
                     }
@@ -386,12 +376,12 @@ namespace AT_SCC
                 // Send data and check if user needs logging
                 foreach (var hexByte in hexBytes)
                 {
-                    await SendASCIIAsync(hexBytes, mySerialPort, logging_check);
+                    await _serialHelp.SendASCIIAsync(hexBytes, mySerialPort, logging_check, textBoxDELAY!, LogFilePath);
 
                     // If also receiving, receive what is sent and output to the panel
                     if (textBoxreceiveType?.Text == "ASCII" || textBoxreceiveType?.Text == "ASCII-HEX")
                     {
-                        ReceiveASCIIAsync(mySerialPort, textBoxesPanel2!, logging_check, textBoxLocationY, textBoxArray, i);
+                        _serialHelp.ReceiveASCIIAsync(mySerialPort, logging_check, textBoxArray, i, LogFilePath);
                     }
                 }
 
@@ -415,136 +405,6 @@ namespace AT_SCC
             if (textBoxSTATUS != null) textBoxSTATUS.Text = "PORT CLOSED";
         }
 
-        private async Task SendStringAsync(SerialPort mySerialPort, string textToSend, bool loggingEnabled, string logFilePath) // task to send strings
-        {
-            mySerialPort.WriteLine(textToSend);
-
-            if (loggingEnabled)
-            {
-                using var logFile = new StreamWriter(logFilePath, true);
-                logFile.WriteLine($"{DateTime.Now:yyyy-MM-dd HH:mm:ss}: [SENT STRING]: {textToSend}\n");
-            }
-
-            await Task.Delay(int.Parse(textBoxDELAY!.Text));
-        }
-
-        private static void ReceiveStringAsync(SerialPort mySerialPort, TextBox receivedTextBox, bool loggingEnabled, string logFilePath)    // task to receive strings
-        {
-            if (mySerialPort.BytesToRead > 0) {
-            receivedTextBox.Text = mySerialPort.ReadLine();
-            if (loggingEnabled)
-            {
-                using var logFile = new StreamWriter(logFilePath, true);
-                logFile.WriteLine($"{DateTime.Now:yyyy-MM-dd HH:mm:ss}: [RECEIVED STRING]: {receivedTextBox.Text}\n");
-            }
-
-            if (receivedTextBox.TextLength >= receivedTextBox.Width * 3)
-            {
-                receivedTextBox.Text = receivedTextBox.Text[(receivedTextBox.TextLength - receivedTextBox.Width * 3)..];
-            }
-
-            }
-            
-        }
-
-        private async Task SendBytesAsync(SerialPort serialPort, string textBoxType, IEnumerable<TextBox> textBoxes) // task to send bytes or byte collections
-        {
-            foreach (var textBox in textBoxes)
-            {
-                if (!string.IsNullOrWhiteSpace(textBox.Text))
-                {
-                    var inputValues = textBox.Text.Split(' ');
-                    var bytesToSend = new List<byte>(); // adds data to the list to send
-
-                    foreach (var value in inputValues)
-                    {
-                        if (byte.TryParse(value, NumberStyles.HexNumber, CultureInfo.InvariantCulture, out var byteValue))
-                        {
-                            bytesToSend.Add(byteValue);
-                        }
-                        else    // if data is unavailable or not correct type, throw error
-                        {
-                            MessageBox.Show($"Error: Unable to parse byte value '{value}'", "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                            return;
-                        }
-
-                        if (logging_check.Checked)
-                        {
-                            var logFilePath = LogFilePath;
-
-                            using var logFile = new StreamWriter(logFilePath, true);
-                            logFile.WriteLine($"{DateTime.Now:yyyy-MM-dd HH:mm:ss}: [SENT BYTE/BYTE COLLECTION]: {byteValue}");
-                        }
-                    }
-
-                    serialPort.Write(bytesToSend.ToArray(), 0, bytesToSend.Count);
-                    await Task.Delay(int.Parse(textBoxDELAY!.Text));
-                }
-            }
-        }
-
-        private void ReceiveBytesAsync(SerialPort mySerialPort, TextBox[]? textBoxArray, int i)     // task to receive bytes or byte collections
-        {
-            if (mySerialPort.BytesToRead > 0) {
-            var bytesReceived = new List<byte>();
-            var receivedText = new StringBuilder();
-            var receivedTextBox = textBoxArray![i];
-
-            // read bytes from port and convert to text
-            var b = (byte)mySerialPort.ReadByte();
-            bytesReceived.Add(b);
-            receivedText.Append(b + " ");
-            
-            receivedTextBox.Text = Convert.ToString(receivedText);
-
-            // create a new textbox to display the received bytes
-
-            if (logging_check.Checked)
-            {
-                var logFilePath = LogFilePath;
-
-                using var logFile = new StreamWriter(logFilePath, true);
-                logFile.WriteLine($"{DateTime.Now:yyyy-MM-dd HH:mm:ss}: [RECEIVED BYTE/BYTE COLLECTION]: {receivedText}");
-            }
-
-            }
-        }
-
-        private async Task SendASCIIAsync(byte[] hexBytes, SerialPort mySerialPort, CheckBox logging_check)     // task to send ASCII or hex values
-        {
-            foreach (var hexByte in hexBytes)
-            {
-                if (logging_check.Checked)
-                {
-                    var logFilePath = LogFilePath;
-
-                    using var logFile = new StreamWriter(logFilePath, true);
-                    logFile.WriteLine($"{DateTime.Now:yyyy-MM-dd HH:mm:ss}: [SENT ASCII/HEX]: {hexByte.ToString("X2")}");
-                }
-
-                mySerialPort.Write(new byte[] { hexByte }, 0, 1);
-
-                await Task.Delay(int.Parse(textBoxDELAY!.Text));
-            }
-        }
-
-        private void ReceiveASCIIAsync(SerialPort mySerialPort, Panel textBoxesPanel2, CheckBox logging_check, int textBoxLocationY, TextBox[] receivedTextBox, int i)
-        {
-
-           
-            receivedTextBox[i].Text = mySerialPort.ReadExisting();
-
-            if (logging_check.Checked)
-            {
-                var logFilePath = LogFilePath;
-
-                using var logFile = new StreamWriter(logFilePath, true);
-                logFile.WriteLine($"{DateTime.Now:yyyy-MM-dd HH:mm:ss}: [RECEIVED ASCII/HEX]: {receivedTextBox[i]!.Text}");
-            }
-            
-        }
-        // task to receive ASCII or hex values
-
         private void OnReload() // function to reload the ports/program
 
         {
@@ -561,6 +421,7 @@ namespace AT_SCC
         {
             InitializeComponent(); // calls the components fuction in the MainDisplay.Designer.cs file
             _displayHelp = new DisplayHelp();
+            _serialHelp = new SerialHelp();
             // DEFINE THE MAIN OVERALL FORM
 
             existingPorts = [.. AvailablePorts];
